POLIOPM-531: cleanup and reorganize files
